# Conflicts:
#	rocket_front/rocket/src/app/app.module.ts

+++ rocket_back/rocket/src/main/java/com/uber/rocket/entity/ride/Review.java
@@ -33,8 +33,9 @@ public class Review {
 
     @ManyToOne
     private Ride ride;
-    public float getRating() {
-        return ((float)(vehicleRating + driverRating))/2;
+
+    public double getRating() {
+        return ((double)(vehicleRating + driverRating))/2;
     }
 
 }
